% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIASEC.R
\docType{data}
\name{RIASEC}
\alias{RIASEC}
\title{Empirical data for the Basic Interest Markers (RIASEC) scale}
\format{
data.frame
}
\usage{
RIASEC
}
\description{
Empirical data for the Basic Interest Markers (RIASEC) scale (5,000 examinees, 93 variables). This scale was originally proposed by Liao et al. (2008). Data was available at https://openpsychometrics.org/_rawdata/. We only included the first 5,000 complete rows. The scale contains 48 items, ordered 5-point categories. Six dimensions: Realistic (items 1-8); Investigative (items 9-16); Artistic (items 17-24), Social (items 24-32), Enterprising (items 33-40) and Conventional (items 41-48).
}
\references{
Liao, H. Y., Armstrong, P. I., & Rounds, J. (2008). Development and initial validation of public domain Basic Interest Markers. \emph{Journal of Vocational Behavior, 73}(1), 159-183.
}
\keyword{datasets}
