% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEAT.R
\docType{data}
\name{SEAT}
\alias{SEAT}
\title{Empirical data for the Self-Efficacy Academic Text (SEAT) scale}
\format{
data.frame
}
\usage{
SEAT
}
\description{
Empirical data for the Self-Efficacy Academic Text (SEAT) scale (418 examinees, 149 variables). This dataset was used in developmental study of the scale by Collado et al. (2023). Data was available at https://osf.io/7bv3a/. The scale contains 8 items, 5-point ordered response scale. Unidimensional. Spanish sample, university students collected online.
}
\references{
Collado, S., Fidalgo, C., Rodríguez-Rey, R., & Sorrel, M. A. (2023). Development and validation of the self-efficacy for writing and defending academic texts scale. \emph{Psicología Educativa, 29}(1), 91-99.
}
\keyword{datasets}
