% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPM.R
\docType{data}
\name{SPM}
\alias{SPM}
\title{Empirical data for Last series of the Standard Progressive Matrices (SPM) scale}
\format{
data.frame
}
\usage{
SPM
}
\description{
Empirical data for Last series of the Standard Progressive Matrices (SPM) scale (499 examinees, 12 variables). This scale was used Myszkowski and Storme (2018). The authors made the data accesible at available at https://data.mendeley.com/datasets/h3yhs5gy3w/1. The scale contains 12 items, multiple-choice items with 8 response options (1 correct). Unidimensional.
}
\references{
Myszkowski, N., & Storme, M. (2018). A snapshot of g? Binary and polytomous item-response theory investigations of the last series of the Standard Progressive Matrices (SPM-LS). \emph{Intelligence, 68}, 109–116.
}
\keyword{datasets}
