% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPER.R
\docType{data}
\name{TEMPER}
\alias{TEMPER}
\title{Empirical data for Fisher Temperament Inventory (TEMPER) scale}
\format{
data.frame
}
\usage{
TEMPER
}
\description{
Empirical data for Fisher Temperament Inventory (TEMPER) scale (4,967 examinees, 113 variables). This scale was proponsed in Brown et al. (2013). Data was available at https://openpsychometrics.org/_rawdata/. The scale contains 56 items, ordered 4-point categories. Four dimensions: Curious/Energetic (items 1-14), Cautious/Compliant with the social norm (items 15-28), Analytical/Hard-headed (items 29 to 42) and Prosocial/Empathetic (items 43 to 56).
}
\references{
Brown, L.L., Acevedo, B., & Fisher, H. E. (2013). Neural Correlates of Four Broad Temperament Dimensions: Testing Predictions for a Novel Construct of Personality. PloS one, 8(11).
}
\keyword{datasets}
